messages cleanup
