+++ app/lib/pages/chat/widgets/user_message.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:friend_private/backend/schema/message.dart';
-import 'package:friend_private/pages/chat/widgets/ai_message.dart';
 import 'package:friend_private/widgets/extensions/string.dart';
 import 'package:friend_private/utils/other/temp.dart';
 
@@ -42,7 +41,6 @@ class HumanMessage extends StatelessWidget {
               ),
             ],
           ),
-          CopyButton(messageText: message.text, isUserMessage: true,),
         ],
       ),
     );
