+++ app/lib/providers/message_provider.dart
@@ -49,6 +49,11 @@ class MessageProvider extends ChangeNotifier {
     notifyListeners();
   }
 
+  void removeLocalMessage(String id) {
+    messages.removeWhere((m) => m.id == id);
+    notifyListeners();
+  }
+
   Future refreshMessages({bool dropdownSelected = false}) async {
     setLoadingMessages(true);
     if (SharedPreferencesUtil().cachedMessages.isNotEmpty) {
